Sigh...fix the warning.
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36651 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a08e25e5c1d42d5b3e8382169c838226f419761d

+++ chrome/browser/browsing_data_local_storage_helper_unittest.cc
@@ -32,7 +32,7 @@ class BrowsingDataLocalStorageHelperTest : public InProcessBrowserTest {
     const FilePath::CharType* kFilesToCreate[] = {
         kTestFile0, kTestFile1, kTestFileInvalid,
     };
-    for (int i = 0; i < arraysize(kFilesToCreate); ++i) {
+    for (size_t i = 0; i < arraysize(kFilesToCreate); ++i) {
       FilePath file_path = storage_path.Append(kFilesToCreate[i]);
       file_util::WriteFile(file_path, NULL, 0);
     }
